Spike: cognee-rs → WebAssembly — Config 1 (logic-only feasibility)
